+++ ts_runtime/src/peer_tracker/peer_db.rs
@@ -545,6 +545,7 @@ mod test {
             peerapi_dns_proxy: false,
             is_wireguard_only: false,
             exit_node_dns_resolvers: vec![],
+            peer_relay: false,
             service_vips: Default::default(),
         }
     }
@@ -831,6 +832,7 @@ mod test {
                     peerapi_dns_proxy: false,
                     is_wireguard_only: false,
                     exit_node_dns_resolvers: vec![],
+                    peer_relay: false,
                     service_vips: Default::default(),
                 }
             })